Bank plugin fixes

+++ runelite-client/src/main/java/net/runelite/client/plugins/grounditems/GroundItemsPlugin.java
@@ -405,7 +405,7 @@ public class GroundItemsPlugin extends Plugin
 		}
 		else
 		{
-			groundItem.setGePrice(itemManager.getItemPrice(realItemId));
+			groundItem.setGePrice((int) itemManager.getItemPrice(realItemId));
 		}
 
 		return groundItem;
